# docs/swagger.yaml
basePath: /
definitions:
  structs.Cart:
    properties:
      created_at: {}
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      selling_price:
        type: integer
      subtotal:
        type: integer
      updated_at: {}
      user_id:
        type: integer
    type: object
...
  structs.Product:
    properties:
      created_at: {}
      id:
        type: integer
      name:
        type: string
      purchase_price:
        type: integer
      selling_price:
        type: integer
      stock:
        type: integer
      updated_at: {}
    type: object
  structs.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Casheex is a simple RESTful API for a casheer system, designed to support
    basic cashier operations in retail or small business environments.
  title: Casheex
  version: "1.0"
paths:
  /api/cart:
    get:
      description: Retrieve all cart items for the authenticated user, including product
        information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's cart
      parameters:
      - description: Cart Data
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/structs.Cart'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Add product to cart
      tags:
      - Cart
  /api/cart/{id}:
    delete:
      description: Remove a product from the authenticated user's cart by cart ID
      parameters:
      - description: Cart ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Remove product from cart
      tags:
      - Cart
  /api/products:
    get:
      description: Retrieve a list of all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Store a new product into the database
      parameters:
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/structs.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      description: Get a product detail by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Find product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/structs.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/structs.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/structs.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.Response'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"